+++ src/Conf.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace LTDBeget\Yiiic;
+
 use yii\helpers\Console;
 
 /**
@@ -12,6 +13,7 @@ class Conf
     const SHOW_HELP_ALWAYS = 'always';
     const SHOW_HELP_NEVER = 'never';
     const SHOW_HELP_ONCE = 'once';
+    const ENTRY_SCRIPT_CURRENT = 'realpath($_SERVER[argv][0])';
 
     /**
      * @var array
@@ -22,7 +24,7 @@ class Conf
     {
         return [
             'entities' => [
-                'apiReflector' => function($options) {
+                'apiReflector' => function ($options) {
                     return new ApiReflector($options['ignore']);
                 }
             ],
@@ -31,6 +33,7 @@ class Conf
                 'prompt' => 'yiiic',
                 'show_help' => self::SHOW_HELP_ONCE,
                 'show_trace' => false,
+                'entry_script' => self::ENTRY_SCRIPT_CURRENT,
                 'commands' => [
                     'context' => 'c',
                     'quit' => 'q',
@@ -38,7 +41,6 @@ class Conf
                 ],
                 'without_context_prefix' => '/',
                 'height_help' => 5,
-                'result_border' => '=',
                 'style' => [
                     'prompt' => [Console::FG_GREEN, Console::BOLD],
                     'welcome' => [Console::FG_YELLOW, Console::BOLD],
@@ -46,10 +48,12 @@ class Conf
                     'notice' => [Console::FG_YELLOW, Console::BOLD],
                     'help' => [
                         'title' => [Console::FG_YELLOW, Console::UNDERLINE],
-                        'scope' => [Console::FG_YELLOW, Console::ITALIC]
+                        'content' => [Console::FG_YELLOW, Console::ITALIC]
                     ],
                     'result' => [
-                        'border' => [Console::FG_CYAN]
+                        'border' => [Console::FG_CYAN],
+                        'content' => [Console::FG_CYAN],
+                        'separator' => '='
                     ],
                     'error' => [Console::FG_RED, Console::BOLD]
                 ]